Added the database connection, so that the annotations are now saved in the database. On application start, the annotation data is loaded from the database for the highlightings. Corrected the algorithm to create and understand the xpath strings. Some other changes.
Added a new activity for viewing the information of one annotation. The database access is now managed by ContentProvider, because sometimes I got exceptions that the db is locked. Now there is only one connection possible to the database. Some bugfixes.

The server authentication works. The new ConnectionManager takes care of it.